refactor(sage-link): rename property 'size' to 'fontSize'
