+++ libs/core/src/components/sage-link/readme.md
@@ -7,11 +7,18 @@ Link is mainly used as navigational element and usually appear within or directl
 
 ## Properties
 
-| Property            | Attribute  | Description                            | Type                   | Default     |
-| ------------------- | ---------- | -------------------------------------- | ---------------------- | ----------- |
-| `external`          | `external` | When enabled, opens link in a new tab. | `boolean`              | `false`     |
-| `href` _(required)_ | `href`     | The URL that the hyperlink points to.  | `string`               | `undefined` |
-| `size`              | `size`     |  The size of the link.                 | `"lg" \| "md" \| "sm"` | `'lg'`      |
+| Property            | Attribute   | Description                                                     | Type                   | Default     |
+| ------------------- | ----------- | --------------------------------------------------------------- | ---------------------- | ----------- |
+| `external`          | `external`  | When enabled, opens link in a new tab.                          | `boolean`              | `false`     |
+| `fontSize`          | `font-size` |  The Font size follows t-shirt model sm: 12px md: 14px lg: 16px | `"lg" \| "md" \| "sm"` | `'lg'`      |
+| `href` _(required)_ | `href`      | The URL that the hyperlink points to.                           | `string`               | `undefined` |
+
+
+## Slots
+
+| Slot | Description                                        |
+| ---- | -------------------------------------------------- |
+|      | Content is placed between the opening closing tags |
 
 
 ----------------------------------------------
